misc: Obsolete xnu-headers, libc-headers, libm-headers, and CarbonHeaders (https://trac.macports.org/ticket/46521)
git-svn-id: https://svn.macports.org/repository/macports/trunk/dports@131699 d073be05-634f-4543-b044-5fe20cf6d1d6

+++ _resources/port1.0/group/obsolete-1.0.tcl
@@ -44,7 +44,11 @@ platforms       darwin
 maintainers     nomaintainer
 supported_archs noarch
 
-description     this port is only a stub and has been made obsolete by ${replaced_by}
+if {[info exists replaced_by]} {
+    description     this port is only a stub and has been made obsolete by ${replaced_by}
+} else {
+    description     this port is only a stub and has been made obsolete without replacement.
+}
 default long_description ${description}
 
 homepage        http://www.macports.org/
@@ -53,7 +57,11 @@ archive_sites
 distfiles
 
 pre-configure {
-    ui_error "${subport} has been made obsolete by the port ${replaced_by}. Please install ${replaced_by} instead."
+    if {[info exists replaced_by]} {
+        ui_error "${subport} has been made obsolete by the port ${replaced_by}. Please install ${replaced_by} instead."
+    } else {
+        ui_error "${subport} is an obsolete port.  Please uninstall it."
+    }
     return -code error "obsolete port"
 }
